add prevent to delete data .

+++ app/Http/Controllers/AmenityController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AmenityRequest;
 use App\Http\Resources\AmenityCollection;
 use App\Http\Resources\AmenityResource;
 use App\Models\Amenity;
+use App\Models\Venue;
 
 class AmenityController extends Controller
 {
@@ -41,6 +42,12 @@ class AmenityController extends Controller
 
     public function destroy(Amenity $amenity)
     {
+        if ($amenity->venues()->exists()) {
+            return response()->json([
+                "message" => "Can't delete amenity because this amenity associate with venues"
+            ], 403);
+        }
+
         $amenity->delete();
 
         return response()->noContent();
